Fix purge script issue

+++ purge-old-files.sh
@@ -4,7 +4,7 @@
 _LOG_FILE=$BISERVER_HOME/tomcat/logs/purge.log
 
 [ -f ${_LOG_FILE}.old ] && rm -f ${_LOG_FILE}.old
-mv $_LOG_FILE ${_LOG_FILE}.old
+[ -f ${_LOG_FILE} ] && mv $_LOG_FILE ${_LOG_FILE}.old
 
 log() {
   [[ "$2" ]] && echo "[`date +'%Y-%m-%d %H:%M:%S.%N'`] - $1 - $2" >> $_LOG_FILE
